% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompositeIndex-package.R, R/CompositeIndex.R
\docType{package}
\name{CompositeIndex}
\alias{CompositeIndex}
\alias{CompositeIndex-package}
\alias{CompositeIndex}
\title{CompositeIndex: return composite economic index}
\usage{
CompositeIndex(consistent, leading = NULL, lagging = NULL, co_wgt = NULL,
  le_wgt = NULL, la_wgt = NULL, rate_method = c("division", "difference",
  "auto"), start = numeric(), end = numeric(), frequency = 12,
  baseyear = start[1], trend_adjust = TRUE, neighbour = 5,
  mincycle = 15, phase = 5, extr_limit = 3.5, end_num = 6,
  mcd_gap = 8)
}
\arguments{
\item{consistent, leading, lagging}{data frame, including consistent indicators, leading indicators, 
lagging indicators respectively .}

\item{co_wgt, le_wgt, la_wgt}{data frame, including weights respectively for \code{consistent}, 
\code{leading}, \code{lagging}.}

\item{rate_method}{calculating method for symmetry rate. \code{auto} means progress selects \code{division}
or \code{difference} automatically .}

\item{start, end}{start and end of time series, it's usage like \code{start=c(2007,2}, \code{end=c(2017,12}.}

\item{frequency}{12 for monthly data , 4 for seasonal data . In addition , \code{start}, \code{end}, and 
\code{frequency} will determine a number of observation,which the actually number of data will be equal to
. Thus data you input will not be permitted to lack some moth or quarter, otherwise , result will be not 
correct .}

\item{baseyear}{final index will base on a certain year(whose indices' mean will be regarded as 100) to 
adjust relative size of index .}

\item{trend_adjust}{\code{logical}, whether to adjust trend of index 
base on compounded average growth, \code{TRUE} : yes ; \code{FALSE }: no.}

\item{neighbour...mcd_gap}{refer to \link{bbturns}, see \url{https://github.com/Nisus-Liu/bbturns}.}
}
\description{
CompositeIndex: return composite economic index

composite consistent index from consistent indicators, composite leading index from leading indicators, composite lagging index from lagging indicators.
The consistent / leading / lagging indicators were determined by corresponding expert 
or algorithms like \link{KL}, \link{crosscor} ... Generally, these are economic indicators,
whose value is itself, or cycle part from seasonal-adjusted (\eqn{C}), or its year-on-year growth rate 
of trend and cycle part (\eqn{TC}) from seasonal-adjusted.
}
\details{
This package depend on package named \code{bbturns} from \emph{GitHub}, you can get it like :
\code{install.packages("devtools")} >> \code{library(devtools)} >> \code{install_github("Nisus-Liu/bbturns")}.
}
\examples{
CI <- CompositeIndex(consistent , leading , co_wgt = consistent_wgt, start = c(2006,1), end = c(2016,7), baseyear = 2006)
}
\author{
Hejun Liu (Nisus-Liu), e-mail: liuhejunlj@163.com
}
