% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{consistent}
\alias{consistent}
\title{economic indicators regarded as consistent indicators}
\format{A 15 x 127 data.frame}
\usage{
consistent
}
\details{
This data set is a format of data.frame, contains monthly  year on year growth of trend and cycle 
(TC) of economic time series, like 'Ind_AV', 'Investment', 'Total_sales' ... from 2006.01 
to 2016.07, 127 observations.
}
\keyword{datasets}
