% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{consistent_wgt}
\alias{consistent_wgt}
\title{weights for dataset \code{consistent}}
\format{A 15 x 1 data.frame}
\usage{
consistent_wgt
}
\details{
This data set is a format of data.frame, contains weights and will weight to 
consistent indicators (\code{consistent}).
}
\keyword{datasets}
