% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_current.R
\name{current2cumulative}
\alias{current2cumulative}
\title{convert current value to cumulative value}
\usage{
current2cumulative(x, cycle = 12, start = 1)
}
\arguments{
\item{x}{a numeric vector or a data frame}

\item{cycle}{12 if monthly ; 4 if seasonal ...}

\item{start}{start month or quarter ...}
}
\value{
A data frame.
}
\description{
sometimes should convert current value to cumulative value if necessity, although using 
current value is common.
}
\details{
No matter what class of \code{x} (vector or data.frame), output value will be a data frame.
If \code{x} is data frame, progress will convert it's all columns to cumulative value.
}
\examples{
a = c(1:10)
b = c(1,3,6,1,8,4,2,9,10,3)
c = c(rep(c(1,2,3),c(2,3,5)))
df = data.frame(a,b,c)
current2cumulative(df,cycle = 5,start = 1 )
current2cumulative(df,cycle = 5,start = 2 )
current2cumulative(df,cycle = 5,start = 7 ) # fail
current2cumulative(df,cycle = 25 )  # when cycle > nobs.
current2cumulative(df,cycle = 50,50 )

}
\seealso{
\link{cumulative2current}
}
