% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KL.R
\name{KL}
\alias{KL}
\title{Kullback-Leibler (KL) information}
\usage{
KL(base, divergent, lag = 12, group_lag = 3)
}
\arguments{
\item{base}{a numeric vector, containing a basic indicator for comparing to other indicators in order to recognize who is leading indicator ...}

\item{divergent}{a data frame containing indicators to be compared with \code{base },
and then to divide to 3 data frame, storing lagging, convenient, lagging indicators respectively.}

\item{lag}{max lag (forward or afterward) for calculating \code{KL information}, at range of \eqn{[-lag, lag]}.}

\item{group_lag}{threshold value of lag for divide different indicator, default is 3, means that 
lagging or leading lag at range of \eqn{[-3, 3]} will group the indicator to consistent indicator.}
}
\value{
A list includes some value as follow :
\describe{
  \item{Coincident_Indicator}{data frame, including coincident indicator}
  \item{Leading_Indicator}{data frame, including leading indicator}
  \item{Lagging_Indicator}{data frame, lagging indicator}
  \item{KL_Matrix}{matrix including KL value and lag information}
  \item{KL_Min}{data frame, including vector of \code{KL_lag} : lag corresponding min KL value and 
  vector of \code{KL_min} : min KL value}
}
}
\description{
A approach to separate leading, consistent, lagging indicators from economic indicators.
}
\examples{
x1 <- c( 5, 9, 5, 12, 13, 8, 14, 14, 8, 10)
x2 <- c( 9, 15, 5, 14, 8, 6, 12, 19,8, 7)
base<-c( 26, 13, 25, 27 ,25 ,13 ,32 ,20 ,23 ,29)
dt<-data.frame(x1, x2, base)
KL(base,dt,lag = 3,group_lag = 1)

}
