% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{leading}
\alias{leading}
\title{electricity consumption indicators regarded as leading indicators}
\format{A 15 x 127 data.frame}
\usage{
leading
}
\details{
This data set is a format of data.frame, contains monthly  year on year growth of trend and cycle 
(TC) of electricity consumption time series, like 'tertiary', 'nonmetal_mining', 'farm_sideline' ... from 2006.01 
to 2016.07, 127 observations.
}
\keyword{datasets}
